+++ emcc.py
@@ -27,7 +27,6 @@ import logging
 import os
 import re
 import shlex
-import shutil
 import sys
 import time
 import tarfile
@@ -78,7 +77,6 @@ EXTRA_INCOMING_JS_API = [
 
 SIMD_INTEL_FEATURE_TOWER = ['-msse', '-msse2', '-msse3', '-mssse3', '-msse4.1', '-msse4.2', '-msse4', '-mavx']
 SIMD_NEON_FLAGS = ['-mfpu=neon']
-COMPILE_ONLY_FLAGS = {'--default-obj-ext'}
 LINK_ONLY_FLAGS = {
     '--bind', '--closure', '--cpuprofiler', '--embed-file',
     '--emit-symbol-map', '--emrun', '--exclude-file', '--extern-post-js',
@@ -149,7 +147,6 @@ class EmccOptions:
     self.memory_init_file = None
     self.use_preload_cache = False
     self.use_preload_plugins = False
-    self.default_object_extension = '.o'
     self.valid_abspaths = []
     # Specifies the line ending format to use for all generated text files.
     # Defaults to using the native EOL on each platform (\r\n on Windows, \n on
@@ -534,6 +531,13 @@ emcc: supported targets: llvm bitcode, WebAssembly, NOT elf
 ''')
     return 0
 
+  ## Process argument and setup the compiler
+  state = EmccState(args)
+  options, newargs = phase_parse_arguments(state)
+
+  if not shared.SKIP_SUBPROCS:
+    shared.check_sanity()
+
   if '--version' in args:
     print(version_string())
     print('''\
@@ -565,13 +569,6 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       parts = [x for x in parts if x not in ['-c', '-o', '-v', '-emit-llvm'] and input_file not in x and temp_target not in x]
       print(shared.shlex_join(parts[1:]))
     return 0
-
-  ## Process argument and setup the compiler
-  state = EmccState(args)
-  options, newargs = phase_parse_arguments(state)
-
-  if not shared.SKIP_SUBPROCS:
-    shared.check_sanity()
 
   if 'EMMAKEN_NO_SDK' in os.environ:
     exit_with_error('EMMAKEN_NO_SDK is no longer supported.  The standard -nostdlib and -nostdinc flags should be used instead')
@@ -618,9 +615,6 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       print(libname)
     return 0
 
-  if not input_files and not state.link_flags:
-    exit_with_error('no input files')
-
   if options.reproduce:
     create_reproduce_file(options.reproduce, args)
 
@@ -641,8 +635,6 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
     return link.run(linker_inputs, options, state, newargs)
   else:
     logger.debug('stopping after compile phase')
-    for flag in state.link_flags:
-      diagnostics.warning('unused-command-line-argument', "argument unused during compilation: '%s'" % flag[1])
     return 0
 
 
@@ -670,9 +662,9 @@ def phase_parse_arguments(state):
   newargs = diagnostics.capture_warnings(newargs)
 
   for i in range(len(newargs)):
-    if newargs[i] in ('-l', '-L', '-I'):
-      # Scan for individual -l/-L/-I arguments and concatenate the next arg on
-      # if there is no suffix
+    if newargs[i] in ('-l', '-L', '-I', '-z'):
+      # Scan for flags that can be written as either one or two arguments
+      # and normalize them to the single argument form.
       newargs[i] += newargs[i + 1]
       newargs[i + 1] = ''
 
@@ -761,30 +753,16 @@ def phase_setup(options, state, newargs):
         else:
           message = arg + ': Unknown format, not a static library!'
         exit_with_error(message)
-      if file_suffix in DYNAMICLIB_ENDINGS and not building.is_bitcode(arg) and not building.is_wasm(arg):
-        # For shared libraries that are neither bitcode nor wasm, assuming its local native
-        # library and attempt to find a library by the same name in our own library path.
-        # TODO(sbc): Do we really need this feature?  See test_other.py:test_local_link
-        libname = removeprefix(get_library_basename(arg), 'lib')
-        flag = '-l' + libname
-        diagnostics.warning('map-unrecognized-libraries', f'unrecognized file type: `{arg}`.  Mapping to `{flag}` and hoping for the best')
-        state.add_link_flag(i, flag)
-      else:
-        input_files.append((i, arg))
+      input_files.append((i, arg))
     elif arg.startswith('-L'):
       state.add_link_flag(i, arg)
-      newargs[i] = ''
     elif arg.startswith('-l'):
       state.add_link_flag(i, arg)
-      newargs[i] = ''
     elif arg == '-z':
       state.add_link_flag(i, newargs[i])
       state.add_link_flag(i + 1, newargs[i + 1])
-      newargs[i] = ''
-      newargs[i + 1] = ''
     elif arg.startswith('-z'):
       state.add_link_flag(i, newargs[i])
-      newargs[i] = ''
     elif arg.startswith('-Wl,'):
       # Multiple comma separated link flags can be specified. Create fake
       # fractional indices for these: -Wl,a,b,c,d at index 4 becomes:
@@ -792,16 +770,10 @@ def phase_setup(options, state, newargs):
       link_flags_to_add = arg.split(',')[1:]
       for flag_index, flag in enumerate(link_flags_to_add):
         state.add_link_flag(i + float(flag_index) / len(link_flags_to_add), flag)
-      newargs[i] = ''
     elif arg == '-Xlinker':
       state.add_link_flag(i + 1, newargs[i + 1])
-      newargs[i] = ''
-      newargs[i + 1] = ''
     elif arg == '-s':
-      # -s and some other compiler flags are normally passed onto the linker
-      # TODO(sbc): Pass this and other flags through when using lld
-      # link_flags.append((i, arg))
-      newargs[i] = ''
+      state.add_link_flag(i, newargs[i])
     elif arg == '-':
       input_files.append((i, arg))
       newargs[i] = ''
@@ -836,23 +808,6 @@ def phase_setup(options, state, newargs):
         diagnostics.warning(
             'unused-command-line-argument',
             "linker flag ignored during compilation: '%s'" % arg)
-    if state.has_dash_c:
-      if '-emit-llvm' in newargs:
-        options.default_object_extension = '.bc'
-    elif state.has_dash_S:
-      if '-emit-llvm' in newargs:
-        options.default_object_extension = '.ll'
-      else:
-        options.default_object_extension = '.s'
-    elif '-M' in newargs or '-MM' in newargs:
-      options.default_object_extension = '.mout' # not bitcode, not js; but just dependency rule of the input file
-
-  else:
-    for arg in state.orig_args:
-      if any(arg.startswith(f) for f in COMPILE_ONLY_FLAGS):
-        diagnostics.warning(
-            'unused-command-line-argument',
-            "compiler flag ignored during linking: '%s'" % arg)
 
   if settings.MAIN_MODULE or settings.SIDE_MODULE:
     settings.RELOCATABLE = 1
@@ -961,11 +916,6 @@ def get_clang_output_extension(state):
 
 @ToolchainProfiler.profile_block('compile inputs')
 def phase_compile_inputs(options, state, newargs, input_files):
-  def is_link_flag(flag):
-    if flag in ('-nostdlib', '-nostartfiles', '-nolibc', '-nodefaultlibs'):
-      return True
-    return flag.startswith(('-l', '-L', '-Wl,'))
-
   if shared.run_via_emxx:
     compiler = [shared.CLANG_CXX]
   else:
@@ -975,7 +925,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
     logger.debug('using compiler wrapper: %s', config.COMPILER_WRAPPER)
     compiler.insert(0, config.COMPILER_WRAPPER)
 
-  compile_args = [a for a in newargs if a and not is_link_flag(a)]
+  compile_args = newargs
   system_libs.ensure_sysroot()
 
   def get_language_mode(args):
@@ -1012,8 +962,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
     # output the dependency rule. Warning: clang and gcc behave differently
     # with -MF! (clang seems to not recognize it)
     logger.debug(('just preprocessor ' if state.has_dash_E else 'just dependencies: ') + ' '.join(cmd))
-    shared.check_call(cmd)
-    return []
+    shared.exec_process(cmd)
 
   # Precompiled headers support
   if state.mode == Mode.PCH:
@@ -1025,8 +974,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
     if options.output_file:
       cmd += ['-o', options.output_file]
     logger.debug(f"running (for precompiled headers): {cmd[0]} {' '.join(cmd[1:])}")
-    shared.check_call(cmd)
-    return []
+    shared.exec_process(cmd)
 
   if state.mode == Mode.COMPILE_ONLY:
     inputs = [i[1] for i in input_files]
@@ -1038,18 +986,17 @@ def phase_compile_inputs(options, state, newargs, input_files):
       cmd += ['-o', options.output_file]
       if get_file_suffix(options.output_file) == '.bc' and not settings.LTO and '-emit-llvm' not in state.orig_args:
         diagnostics.warning('emcc', '.bc output file suffix used without -flto or -emit-llvm.  Consider using .o extension since emcc will output an object file, not a bitcode file')
-    shared.check_call(cmd)
-    if not options.output_file:
-      # Rename object files to match --default-obj-ext
-      # TODO: Remove '--default-obj-ext' to reduce this complexity
-      ext = get_clang_output_extension(state)
-      if options.default_object_extension != ext:
-        for i in inputs:
-          output = unsuffixed_basename(i) + ext
-          new_output = unsuffixed_basename(i) + options.default_object_extension
-          shutil.move(output, new_output)
-    return []
+    shared.exec_process(cmd)
 
+  # In COMPILE_AND_LINK we need to compile source files too, but we also need to
+  # filter out the link flags
+
+  def is_link_flag(flag):
+    if flag in ('-nostdlib', '-nostartfiles', '-nolibc', '-nodefaultlibs', '-s'):
+      return True
+    return flag.startswith(('-l', '-L', '-Wl,', '-z'))
+
+  compile_args = [a for a in compile_args if a and not is_link_flag(a)]
   linker_inputs = []
   seen_names = {}
 
@@ -1059,7 +1006,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
     return unsuffixed(name) + '_' + seen_names[name] + shared.suffix(name)
 
   def get_object_filename(input_file):
-    return in_temp(unsuffixed(uniquename(input_file)) + options.default_object_extension)
+    return in_temp(shared.replace_suffix(uniquename(input_file), '.o'))
 
   def compile_source_file(i, input_file):
     logger.debug(f'compiling source file: {input_file}')
@@ -1404,9 +1351,7 @@ def parse_args(newargs):
     elif arg == '-fignore-exceptions':
       settings.DISABLE_EXCEPTION_CATCHING = 1
     elif check_arg('--default-obj-ext'):
-      options.default_object_extension = consume_arg()
-      if not options.default_object_extension.startswith('.'):
-        options.default_object_extension = '.' + options.default_object_extension
+      exit_with_error('--default-obj-ext is no longer supported by emcc')
     elif arg.startswith('-fsanitize=cfi'):
       exit_with_error('emscripten does not currently support -fsanitize=cfi')
     elif check_arg('--output_eol'):